// src/app/shared/interfaces/orderInterface.ts
export interface Order {
  cart: Product[];
  coupon: string | null;
}

export interface Product {
  product_id: number;
  quantity: number;
}

// Generated by https://quicktype.io

export interface PaySkyConfig {
  MID: number;
  TID: number;
  SecureHash: string;
  MerchantReference: number;
  amount_egp?: number;
  amount_usd?: number;
  AmountTrxn: number;
  currency: string;
  CustomerMobile: string;
  orderId: number;
}